# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DiceGame",
        "description": "URI's List for DiceGameApi",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/v1/players/{id}/play": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Play a game",
                "description": "Play method",
                "operationId": "play",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the player",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Game stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Game stored successfully"
                                        },
                                        "game": {
                                            "$ref": "#/components/schemas/PlayGame"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PlayGame": {
                "required": [
                    "user_id",
                    "dice1_value",
                    "dice2_value",
                    "sum",
                    "result"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer"
                    },
                    "dice1_value": {
                        "type": "integer"
                    },
                    "dice2_value": {
                        "type": "integer"
                    },
                    "sum": {
                        "type": "integer"
                    },
                    "result": {
                        "type": "string",
                        "enum": [
                            "won",
                            "lost"
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    }
}